Migrate releases to feature module (PSY-99)
Move release components, hooks, admin UI, and types into
features/releases/ with co-located structure. Update all imports.

Closes PSY-99
