+++ frontend/app/admin/releases/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { ReleaseManagement } from '@/app/admin/releases/_components/ReleaseManagement'
+import { ReleaseManagement } from '@/features/releases/admin/ReleaseManagement'
 
 export default function AdminReleasesPage() {
   return <ReleaseManagement />
