+++ frontend/components/artists/ArtistDetail.tsx
@@ -17,7 +17,7 @@ import {
 } from 'lucide-react'
 import { useQueryClient } from '@tanstack/react-query'
 import { useArtist } from '@/lib/hooks/artists/useArtists'
-import { useArtistReleases } from '@/lib/hooks/releases/useReleases'
+import { useArtistReleases } from '@/features/releases/hooks/useReleases'
 import { useArtistLabels, useLabelRoster } from '@/lib/hooks/labels/useLabels'
 import { queryKeys } from '@/lib/queryClient'
 import { useIsAuthenticated } from '@/lib/hooks/auth/useAuth'
@@ -38,8 +38,8 @@ import { Input } from '@/components/ui/input'
 import { Alert, AlertDescription } from '@/components/ui/alert'
 import { TabsContent } from '@/components/ui/tabs'
 import { Badge } from '@/components/ui/badge'
-import { getReleaseTypeLabel } from '@/lib/types/release'
-import type { ArtistReleaseListItem } from '@/lib/types/release'
+import { getReleaseTypeLabel } from '@/features/releases/types'
+import type { ArtistReleaseListItem } from '@/features/releases/types'
 import type { ArtistLabel } from '@/lib/types/label'
 
 interface ArtistDetailProps {
