+++ frontend/features/releases/admin/index.ts
@@ -1,0 +1,1 @@
+export { ReleaseManagement } from './ReleaseManagement'
