+++ frontend/features/releases/hooks/index.ts
@@ -1,0 +1,17 @@
+export {
+  useReleases,
+  useRelease,
+  useArtistReleases,
+} from './useReleases'
+
+export {
+  type CreateReleaseArtistInput,
+  type CreateReleaseLinkInput,
+  type CreateReleaseInput,
+  type UpdateReleaseInput,
+  useCreateRelease,
+  useUpdateRelease,
+  useDeleteRelease,
+  useAddReleaseLink,
+  useRemoveReleaseLink,
+} from './useAdminReleases'
