+++ frontend/lib/hooks/index.ts
@@ -6,7 +6,6 @@ export * from './admin'
 export * from './shows'
 export * from './artists'
 export * from './venues'
-export * from './releases'
 export * from './labels'
 export * from './festivals'
 export * from './auth'
