--- frontend/lib/hooks/releases/index.ts
@@ -1,5 +1,0 @@
-export {
-  useReleases,
-  useRelease,
-  useArtistReleases,
-} from './useReleases'
